Added NavigatorSyncHandlers.
